Add download module
+ Add a new android library (module) called download where the data download will happen
+ Refactor (remove mobile folder in app module)
