+++ app/src/main/java/com/chjaeggi/boardgametracker/about/AboutViewModel.kt
@@ -1,0 +1,7 @@
+package com.chjaeggi.boardgametracker.about
+
+import androidx.lifecycle.ViewModel
+
+class AboutViewModel : ViewModel() {
+    // TODO: Implement the ViewModel
+}
