+++ app/src/main/java/com/chjaeggi/boardgametracker/details/DetailsViewModel.kt
@@ -1,0 +1,12 @@
+package com.chjaeggi.boardgametracker.details
+
+import androidx.lifecycle.ViewModel
+import com.chjaeggi.boardgametracker.download.DownloadManager
+import timber.log.Timber
+
+class DetailsViewModel(downloadManager: DownloadManager) : ViewModel() {
+
+    init {
+        Timber.d(downloadManager.getDoc())
+    }
+}
