+++ app/src/main/java/com/chjaeggi/boardgametracker/home/HomeFragment.kt
@@ -1,0 +1,34 @@
+package com.chjaeggi.boardgametracker.home
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.databinding.DataBindingUtil
+import androidx.fragment.app.Fragment
+import androidx.navigation.Navigation
+import androidx.navigation.ui.setupWithNavController
+import com.chjaeggi.boardgametracker.R
+import com.chjaeggi.boardgametracker.databinding.FragmentHomeBinding
+
+class HomeFragment : Fragment() {
+
+    private lateinit var binding: FragmentHomeBinding
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        binding = DataBindingUtil.inflate(inflater, R.layout.fragment_home, container, false)
+        binding.setLifecycleOwner(this)
+        return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        val bottomNavController =
+            Navigation.findNavController(requireActivity(), R.id.home_nav_fragment)
+        binding.homeNavBar.setupWithNavController(bottomNavController)
+    }
+
+}
