+++ app/src/main/java/com/chjaeggi/boardgametracker/home/favorites/FavoritesViewModel.kt
@@ -1,0 +1,7 @@
+package com.chjaeggi.boardgametracker.home.favorites
+
+import androidx.lifecycle.ViewModel
+
+class FavoritesViewModel : ViewModel() {
+    // TODO: Implement the ViewModel
+}
