+++ app/src/main/java/com/chjaeggi/boardgametracker/home/overview/ListItemGame.kt
@@ -1,0 +1,11 @@
+package com.chjaeggi.boardgametracker.home.overview
+
+data class ListItemGame(
+    val thumbnailUrl: String,
+    val name: String,
+    val minPlayers: Int,
+    val maxPlayers: Int,
+    val minDuration: Int,
+    val maxDuration: Int,
+    val isFavorite: Boolean
+)
