+++ app/src/main/java/com/chjaeggi/boardgametracker/home/statistics/StatisticsViewModel.kt
@@ -1,0 +1,7 @@
+package com.chjaeggi.boardgametracker.home.statistics
+
+import androidx.lifecycle.ViewModel
+
+class StatisticsViewModel : ViewModel() {
+    // TODO: Implement the ViewModel
+}
