--- app/src/main/java/com/chjaeggi/boardgametracker/mobile/about/AboutViewModel.kt
@@ -1,7 +1,0 @@
-package com.chjaeggi.boardgametracker.mobile.about
-
-import androidx.lifecycle.ViewModel
-
-class AboutViewModel : ViewModel() {
-    // TODO: Implement the ViewModel
-}
