--- app/src/main/java/com/chjaeggi/boardgametracker/mobile/details/DetailsViewModel.kt
@@ -1,8 +1,0 @@
-package com.chjaeggi.boardgametracker.mobile.details
-
-import androidx.lifecycle.ViewModel
-
-class DetailsViewModel : ViewModel() {
-
-
-}
