--- app/src/main/java/com/chjaeggi/boardgametracker/mobile/home/favorites/FavoritesViewModel.kt
@@ -1,7 +1,0 @@
-package com.chjaeggi.boardgametracker.mobile.home.favorites
-
-import androidx.lifecycle.ViewModel
-
-class FavoritesViewModel : ViewModel() {
-    // TODO: Implement the ViewModel
-}
