--- app/src/main/java/com/chjaeggi/boardgametracker/mobile/home/statistics/StatisticsViewModel.kt
@@ -1,7 +1,0 @@
-package com.chjaeggi.boardgametracker.mobile.home.statistics
-
-import androidx.lifecycle.ViewModel
-
-class StatisticsViewModel : ViewModel() {
-    // TODO: Implement the ViewModel
-}
