--- app/src/main/java/com/chjaeggi/boardgametracker/mobile/util/SimpleAnimationListener.kt
@@ -1,13 +1,0 @@
-package com.chjaeggi.boardgametracker.mobile.util
-
-import android.animation.Animator
-
-open class SimpleAnimatorListener : Animator.AnimatorListener {
-    override fun onAnimationRepeat(animator: Animator) {}
-
-    override fun onAnimationEnd(animator: Animator) {}
-
-    override fun onAnimationCancel(animator: Animator) {}
-
-    override fun onAnimationStart(animator: Animator) {}
-}
