+++ app/src/main/java/com/chjaeggi/boardgametracker/util/ImageViewBindings.kt
@@ -1,0 +1,10 @@
+package com.chjaeggi.boardgametracker.util
+
+import android.widget.ImageView
+import androidx.databinding.BindingAdapter
+import com.bumptech.glide.Glide
+
+@BindingAdapter("imageUrl")
+fun ImageView.setImageUrl(url: String?) {
+    if (url != null) Glide.with(context).load(url).into(this)
+}
