+++ app/src/main/java/com/chjaeggi/boardgametracker/util/SimpleAnimationListener.kt
@@ -1,0 +1,13 @@
+package com.chjaeggi.boardgametracker.util
+
+import android.animation.Animator
+
+open class SimpleAnimatorListener : Animator.AnimatorListener {
+    override fun onAnimationRepeat(animator: Animator) {}
+
+    override fun onAnimationEnd(animator: Animator) {}
+
+    override fun onAnimationCancel(animator: Animator) {}
+
+    override fun onAnimationStart(animator: Animator) {}
+}
