+++ app/src/main/res/layout/activity_details.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="model"
-            type="com.chjaeggi.boardgametracker.mobile.details.DetailsViewModel"/>
+            type="com.chjaeggi.boardgametracker.details.DetailsViewModel"/>
 
     </data>
