+++ app/src/main/res/layout/fragment_about.xml
@@ -9,7 +9,7 @@
     <FrameLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".mobile.about.AboutFragment">
+        tools:context=".about.AboutFragment">
 
         <TextView
             android:layout_width="wrap_content"
