+++ app/src/main/res/layout/fragment_overview.xml
@@ -2,7 +2,7 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".mobile.home.overview.OverviewFragment">
+    tools:context=".home.overview.OverviewFragment">
 
     <ScrollView
     android:layout_width="match_parent"
