+++ app/src/main/res/navigation/nav_main.xml
@@ -7,12 +7,12 @@
 
     <fragment
         android:id="@+id/home_fragment"
-        android:name="com.chjaeggi.boardgametracker.mobile.home.HomeFragment"
+        android:name="com.chjaeggi.boardgametracker.home.HomeFragment"
         android:label=" "
         tools:layout="@layout/fragment_home" />
     <fragment
         android:id="@+id/about_fragment"
-        android:name="com.chjaeggi.boardgametracker.mobile.about.AboutFragment"
+        android:name="com.chjaeggi.boardgametracker.about.AboutFragment"
         android:label="About"
         tools:layout="@layout/fragment_about" />
