+++ download/src/main/java/com/chjaeggi/boardgametracker/download/DownloadManager.kt
@@ -1,0 +1,14 @@
+package com.chjaeggi.boardgametracker.download
+
+import org.jsoup.Jsoup
+import org.jsoup.nodes.Document
+
+class DownloadManager {
+
+    private var doc: Document = Jsoup.connect("https://boardgamegeek.com/browse/boardgame").get()
+
+    fun getDoc(): String {
+        return doc.html()
+    }
+
+}
